refactor: 검색어 서비스 위치 변경
