+++ src/main/java/com/recipe/app/src/recipe/api/SearchKeywordController.java
@@ -1,10 +1,9 @@
 package com.recipe.app.src.recipe.api;
 
 import com.recipe.app.common.response.BaseResponse;
-import com.recipe.app.src.recipe.application.SearchKeywordService;
+import com.recipe.app.src.recipe.application.keyword.SearchKeywordService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
